moved App to src directory
